Add reverbia-pages.css for dashboard and calendar styles
- Introduced styles for dashboard components including token cards, donut charts, and token stats.
- Added calendar styles with responsive design for calendar cards, headers, and day indicators.
- Implemented slider styles for dynamic content display.
- Enhanced button and link styles for better user interaction.
- Established a consistent color scheme using CSS variables for theming.

+++ database/migrations/0001_01_01_000000_create_users_table.php
@@ -25,9 +25,9 @@ return new class extends Migration
         });
 
         Schema::create('users', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->unsignedInteger('branch_id')->nullable();
-            $table->unsignedInteger('seller_id')->nullable();
+            $table->unsignedBigInteger('seller_id')->nullable();
             $table->string('name');
             $table->string('email')->unique();
             $table->string('phone', 50)->nullable();
